Add files via upload
Python code for investigating the Lorenz system using both Euler's method and order 4 Runge-Kutta in order to develop two approximations. Tools for analyzing the differences in the two approximations (for instance an output of the maximum and minimum Euclidean distance between the two solutions) are printed. Graphical plots versus time and in 3D are displayed. Variable step-size, solution size, and adjustable parameters (initial conditions as well as adjustable Prandtl number and Rayleigh number) are allowed in the function call for greater flexibility in experimenting with values